requirements.txt
python-dotenv==1.0.0
replicate==0.22.0
reportlab==4.0.7
feedparser==6.0.10
requests==2.31.0
Pillow==10.1.0
beautifulsoup4==4.12.2 